Move device.py and harp_serial.py to communication folder
