+++ docs/examples/read_and_write_from_registers/read_and_write_from_registers.py
@@ -1,7 +1,7 @@
 from serial import SerialException
 
+from pyharp.devices.device import Device
 from pyharp.protocol import MessageType, PayloadType
-from pyharp.protocol.device import Device
 from pyharp.protocol.messages import HarpMessage
 
 SERIAL_PORT = (
