+++ pyharp/communication/__init__.py
@@ -1,0 +1,2 @@
+from .device import Device as Device
+from .harp_serial import HarpSerial as HarpSerial
