+++ pyharp/communication/harp_serial.py
@@ -10,18 +10,6 @@ import serial.threaded
 from pyharp.protocol.messages import HarpMessage, MessageType
 
 
-class HarpSerialProtocolOld(serial.threaded.Protocol):
-    # Old implementation (per-byte queue)
-    def __init__(self, read_q: queue.Queue, *args, **kwargs):
-        self._read_q = read_q
-        super().__init__(*args, **kwargs)
-
-    def data_received(self, data: bytes) -> None:
-        for byte in data:
-            self._read_q.put(byte)
-        return super().data_received(data)
-
-
 class HarpSerialProtocol(serial.threaded.Protocol):
     """
     The `HarpSerialProtocol` class deals with the data received from the serial communication.
@@ -66,7 +54,7 @@ class HarpSerialProtocol(serial.threaded.Protocol):
                 # not enough data to read the message type and length
                 break
 
-            message_type = self._buffer[0]
+            # Read length (we can ignore the message type)
             message_length = self._buffer[1]
             total_length = 2 + message_length
             if len(self._buffer) < total_length:
@@ -103,14 +91,12 @@ class HarpSerial:
     msg_q: queue.Queue
     event_q: queue.Queue
 
-    def __init__(self, serial_port: str, use_buffered_protocol: bool = True, **kwargs):
+    def __init__(self, serial_port: str, **kwargs):
         """
         Parameters
         ----------
         serial_port : str
             the serial port used to establish the connection with the Harp device. It must be denoted as `/dev/ttyUSBx` in Linux and `COMx` in Windows, where `x` is the number of the serial port
-        use_buffered_protocol : bool
-            whether to use the buffered protocol for reading data
         """
         # Connect to the Harp device
         self._ser = serial.Serial(serial_port, **kwargs)
@@ -123,28 +109,16 @@ class HarpSerial:
         self.event_q = queue.Queue()
 
         # Start the thread with the `HarpSerialProtocol`
-        self.use_buffered_protocol = use_buffered_protocol
-        protocol_cls = (
-            HarpSerialProtocol if use_buffered_protocol else HarpSerialProtocolOld
-        )
-
         self._reader = serial.threaded.ReaderThread(
             self._ser,
-            partial(protocol_cls, self._read_q),
+            partial(HarpSerialProtocol, self._read_q),
         )
         self._reader.start()
         self._reader.connect()
 
-        # Choose parsing method based on protocol
-        parse_target = (
-            self.parse_harp_msgs_threaded_buffered
-            if use_buffered_protocol
-            else self.parse_harp_msgs_threaded_per_byte
-        )
-
         # Start the thread that parses and separates the events from the remaining messages
         self._parse_thread = threading.Thread(
-            target=parse_target,
+            target=self.parse_harp_msgs_threaded_buffered,
             daemon=True,
         )
         self._parse_thread.start()
@@ -177,30 +151,3 @@ class HarpSerial:
             except Exception as e:
                 self.log.error(f"Error parsing message: {e}")
                 self.log.debug(f"Raw data: {frame}")
-
-    def parse_harp_msgs_threaded_per_byte(self):
-        """
-        Parses the Harp messages and separates the events from the remaining messages.
-        """
-        while True:
-            # Gets the Harp message bytes based on the length byte of the message
-            message_type = self._read_q.get(1)
-            message_length = self._read_q.get(1)
-            message_content = bytes([self._read_q.get() for _ in range(message_length)])
-            self.log.debug(f"reply (type): {message_type}")
-            self.log.debug(f"reply (length): {message_length}")
-            self.log.debug(f"reply (payload): {message_content}")
-
-            # Reconstructs the message into a bytearray
-            frame = bytearray()
-            frame.append(message_type)
-            frame.append(message_length)
-            frame += message_content
-            # Parses the bytearray into a ReplyHarpMessage object
-            msg = HarpMessage.parse(frame)
-
-            # Puts the parsed Harp message into the correct queue
-            if msg.message_type == MessageType.EVENT:
-                self.event_q.put(msg)
-            else:
-                self.msg_q.put(msg)
